#0: Cleanup and unify single/multi device files for Resnet Add N300 performance numbers for Resnet to README
