+++ models/demos/t3000/resnet50/tests/test_perf_e2e_resnet50.py
@@ -4,7 +4,7 @@
 
 import pytest
 
-from models.demos.ttnn_resnet.tests.multi_device.perf_e2e_resnet50 import run_perf_resnet
+from models.demos.ttnn_resnet.tests.perf_e2e_resnet50 import run_perf_resnet
 from models.utility_functions import run_for_wormhole_b0
