+++ models/demos/tg/resnet50/tests/test_resnet50_performant.py
@@ -7,7 +7,7 @@ import ttnn
 from models.utility_functions import (
     run_for_wormhole_b0,
 )
-from models.demos.ttnn_resnet.tests.multi_device.resnet50_performant import (
+from models.demos.ttnn_resnet.tests.resnet50_performant import (
     run_resnet50_inference,
     run_resnet50_2cqs_inference,
     run_resnet50_trace_inference,
