--- models/demos/ttnn_resnet/tests/multi_device/perf_e2e_resnet50.py
@@ -1,463 +1,0 @@
-# SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
-
-# SPDX-License-Identifier: Apache-2.0
-
-import torch
-from loguru import logger
-from transformers import AutoImageProcessor
-import pytest
-import ttnn
-from ttnn.model_preprocessing import (
-    preprocess_model_parameters,
-)
-
-from models.utility_functions import (
-    profiler,
-    disable_persistent_kernel_cache,
-)
-
-from models.demos.ttnn_resnet.tests.resnet50_test_infra import create_test_infra
-
-from models.perf.perf_utils import prep_perf_report
-
-from models.demos.ttnn_resnet.tests.resnet50_test_infra import load_resnet50_model
-from models.demos.ttnn_resnet.tt.custom_preprocessing import create_custom_mesh_preprocessor
-
-try:
-    from tracy import signpost
-
-    use_signpost = True
-except ModuleNotFoundError:
-    use_signpost = False
-
-
-def buffer_address(tensor):
-    addr = []
-    for ten in ttnn.get_device_tensors(tensor):
-        addr.append(ten.buffer_address())
-    return addr
-
-
-def dump_device_profiler(device):
-    if isinstance(device, ttnn.Device):
-        ttnn.DumpDeviceProfiler(device)
-    else:
-        for dev in device.get_device_ids():
-            ttnn.DumpDeviceProfiler(device.get_device(dev))
-
-
-# TODO: Create ttnn apis for this
-ttnn.dump_device_profiler = dump_device_profiler
-
-model_config = {
-    "MATH_FIDELITY": ttnn.MathFidelity.LoFi,
-    "WEIGHTS_DTYPE": ttnn.bfloat8_b,
-    "ACTIVATIONS_DTYPE": ttnn.bfloat8_b,
-}
-
-# TODO: Create ttnn apis for this
-ttnn.buffer_address = buffer_address
-
-
-def run_model(
-    device, tt_inputs, test_infra, mesh_mapper, mesh_composer, num_warmup_iterations, num_measurement_iterations
-):
-    ops_parallel_config = {}
-    tt_inputs_host, input_mem_config = test_infra.setup_l1_sharded_input(
-        device,
-        mesh_mapper=mesh_mapper,
-    )
-    profiler.start("compile")
-    test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("compile")
-    ttnn.dump_device_profiler(device)
-
-    profiler.start("cache")
-    test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("cache")
-    ttnn.dump_device_profiler(device)
-
-    for iter in range(0, num_warmup_iterations):
-        test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-        _ = ttnn.from_device(test_infra.run(), blocking=True)
-        ttnn.dump_device_profiler(device)
-
-    ttnn.synchronize_devices(device)
-    if use_signpost:
-        signpost(header="start")
-    outputs = []
-    profiler.start(f"run")
-    for iter in range(0, num_measurement_iterations):
-        test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-        outputs.append(ttnn.from_device(test_infra.run(), blocking=False))
-    ttnn.synchronize_devices(device)
-    profiler.end(f"run")
-    if use_signpost:
-        signpost(header="stop")
-    ttnn.dump_device_profiler(device)
-
-
-def run_2cq_model(
-    device, tt_inputs, test_infra, mesh_mapper, mesh_composer, num_warmup_iterations, num_measurement_iterations
-):
-    ops_parallel_config = {}
-    tt_inputs_host, sharded_mem_config_DRAM, input_mem_config = test_infra.setup_dram_sharded_input(
-        device,
-        mesh_mapper=mesh_mapper,
-    )
-    tt_image_res = tt_inputs_host.to(device, sharded_mem_config_DRAM)
-    op_event = ttnn.create_event(device)
-    write_event = ttnn.create_event(device)
-    # Initialize the op event so we can write
-    ttnn.record_event(0, op_event)
-
-    profiler.start("compile")
-    ttnn.wait_for_event(1, op_event)
-    ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-    ttnn.record_event(1, write_event)
-    ttnn.wait_for_event(0, write_event)
-    test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-    ttnn.record_event(0, op_event)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("compile")
-    ttnn.dump_device_profiler(device)
-
-    profiler.start("cache")
-    ttnn.wait_for_event(1, op_event)
-    ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-    ttnn.record_event(1, write_event)
-    ttnn.wait_for_event(0, write_event)
-    test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-    ttnn.record_event(0, op_event)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("cache")
-    ttnn.dump_device_profiler(device)
-
-    for iter in range(0, num_warmup_iterations):
-        ttnn.wait_for_event(1, op_event)
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-        ttnn.record_event(1, write_event)
-        ttnn.wait_for_event(0, write_event)
-        test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-        ttnn.record_event(0, op_event)
-        _ = ttnn.from_device(test_infra.run(), blocking=True)
-        ttnn.dump_device_profiler(device)
-
-    ttnn.synchronize_devices(device)
-    if use_signpost:
-        signpost(header="start")
-    outputs = []
-    profiler.start(f"run")
-    for iter in range(0, num_measurement_iterations):
-        ttnn.wait_for_event(1, op_event)
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-        ttnn.record_event(1, write_event)
-        ttnn.wait_for_event(0, write_event)
-        test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-        ttnn.record_event(0, op_event)
-        outputs.append(ttnn.from_device(test_infra.run(), blocking=False))
-    ttnn.synchronize_devices(device)
-    profiler.end(f"run")
-    if use_signpost:
-        signpost(header="stop")
-    ttnn.dump_device_profiler(device)
-
-
-def run_trace_model(
-    device, tt_inputs, test_infra, mesh_mapper, mesh_composer, num_warmup_iterations, num_measurement_iterations
-):
-    ops_parallel_config = {}
-    tt_inputs_host, input_mem_config = test_infra.setup_l1_sharded_input(
-        device,
-        mesh_mapper=mesh_mapper,
-    )
-    # Compile
-    profiler.start("compile")
-    test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-    shape = test_infra.input_tensor.shape
-    dtype = test_infra.input_tensor.dtype
-    layout = test_infra.input_tensor.layout
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("compile")
-    ttnn.dump_device_profiler(device)
-    test_infra.output_tensor.deallocate(force=True)
-
-    profiler.start("cache")
-    test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("cache")
-    ttnn.dump_device_profiler(device)
-
-    # Capture
-    test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
-    test_infra.output_tensor.deallocate(force=True)
-    trace_input_addr = ttnn.buffer_address(test_infra.input_tensor)
-    tid = ttnn.begin_trace_capture(device, cq_id=0)
-    tt_output_res = test_infra.run()
-    tt_image_res = ttnn.allocate_tensor_on_device(
-        shape,
-        dtype,
-        layout,
-        device,
-        input_mem_config,
-    )
-    ttnn.end_trace_capture(device, tid, cq_id=0)
-    assert trace_input_addr == ttnn.buffer_address(tt_image_res)
-    ttnn.dump_device_profiler(device)
-
-    for iter in range(0, num_warmup_iterations):
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res)
-        ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
-        _ = ttnn.from_device(tt_output_res, blocking=True)
-        ttnn.dump_device_profiler(device)
-
-    ttnn.synchronize_devices(device)
-    if use_signpost:
-        signpost(header="start")
-    outputs = []
-    profiler.start(f"run")
-    for iter in range(0, num_measurement_iterations):
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res)
-        ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
-        outputs.append(ttnn.from_device(tt_output_res, blocking=False))
-    ttnn.synchronize_devices(device)
-    profiler.end(f"run")
-    if use_signpost:
-        signpost(header="stop")
-    ttnn.dump_device_profiler(device)
-
-    ttnn.release_trace(device, tid)
-
-
-def run_trace_2cq_model(
-    device, tt_inputs, test_infra, mesh_mapper, mesh_composer, num_warmup_iterations, num_measurement_iterations
-):
-    ops_parallel_config = {}
-    tt_inputs_host, sharded_mem_config_DRAM, input_mem_config = test_infra.setup_dram_sharded_input(
-        device,
-        mesh_mapper=mesh_mapper,
-    )
-    tt_image_res = tt_inputs_host.to(device, sharded_mem_config_DRAM)
-
-    op_event = ttnn.create_event(device)
-    write_event = ttnn.create_event(device)
-    # Initialize the op event so we can write
-    ttnn.record_event(0, op_event)
-
-    profiler.start("compile")
-    ttnn.wait_for_event(1, op_event)
-    ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-    ttnn.record_event(1, write_event)
-    ttnn.wait_for_event(0, write_event)
-    test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-    shape = test_infra.input_tensor.shape
-    dtype = test_infra.input_tensor.dtype
-    layout = test_infra.input_tensor.layout
-    ttnn.record_event(0, op_event)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("compile")
-    ttnn.dump_device_profiler(device)
-
-    profiler.start("cache")
-    ttnn.wait_for_event(1, op_event)
-    ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-    ttnn.record_event(1, write_event)
-    ttnn.wait_for_event(0, write_event)
-    test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-    ttnn.record_event(0, op_event)
-    # Deallocate the previous output tensor here to make allocation match capture setup
-    # This allows us to allocate the input tensor after at the same address
-    test_infra.output_tensor.deallocate(force=True)
-    _ = ttnn.from_device(test_infra.run(), blocking=True)
-    profiler.end("cache")
-    ttnn.dump_device_profiler(device)
-
-    # Capture
-    ttnn.wait_for_event(1, op_event)
-    ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-    ttnn.record_event(1, write_event)
-    ttnn.wait_for_event(0, write_event)
-    test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
-    ttnn.record_event(0, op_event)
-    test_infra.output_tensor.deallocate(force=True)
-    trace_input_addr = ttnn.buffer_address(test_infra.input_tensor)
-
-    tid = ttnn.begin_trace_capture(device, cq_id=0)
-    tt_output_res = test_infra.run()
-    input_tensor = ttnn.allocate_tensor_on_device(
-        shape,
-        dtype,
-        layout,
-        device,
-        input_mem_config,
-    )
-    ttnn.end_trace_capture(device, tid, cq_id=0)
-    assert trace_input_addr == ttnn.buffer_address(input_tensor)
-    ttnn.dump_device_profiler(device)
-
-    for iter in range(0, num_warmup_iterations):
-        ttnn.wait_for_event(1, op_event)
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-        ttnn.record_event(1, write_event)
-        ttnn.wait_for_event(0, write_event)
-        input_tensor = ttnn.reshard(tt_image_res, input_mem_config, input_tensor)
-        ttnn.record_event(0, op_event)
-        ttnn.execute_trace(device, tid, cq_id=0, blocking=True)
-        ttnn.dump_device_profiler(device)
-
-    ttnn.synchronize_devices(device)
-    if use_signpost:
-        signpost(header="start")
-    outputs = []
-    profiler.start(f"run")
-    for iter in range(0, num_measurement_iterations):
-        ttnn.wait_for_event(1, op_event)
-        ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
-        ttnn.record_event(1, write_event)
-        ttnn.wait_for_event(0, write_event)
-        # TODO: Add in place support to ttnn to_memory_config
-        input_tensor = ttnn.reshard(tt_image_res, input_mem_config, input_tensor)
-        ttnn.record_event(0, op_event)
-        ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
-        outputs.append(tt_output_res.cpu(blocking=False))
-    ttnn.synchronize_devices(device)
-    profiler.end(f"run")
-    if use_signpost:
-        signpost(header="stop")
-    ttnn.dump_device_profiler(device)
-
-    ttnn.release_trace(device, tid)
-
-
-def run_perf_resnet(
-    device_batch_size,
-    expected_inference_time,
-    expected_compile_time,
-    hf_cat_image_sample_input,
-    device,
-    model_version,
-    model_location_generator,
-):
-    profiler.clear()
-    disable_persistent_kernel_cache()
-    if device_batch_size <= 2:
-        pytest.skip("Batch size 1 and 2 are not supported with sharded data")
-    num_devices = 1 if isinstance(device, ttnn.Device) else device.get_num_devices()
-    batch_size = device_batch_size * num_devices
-    first_key = f"first_iter_batchsize{batch_size}"
-    second_key = f"second_iter_batchsize{batch_size}"
-    cpu_key = f"ref_key_batchsize{batch_size}"
-    model_name = "microsoft/resnet-50"
-
-    image = hf_cat_image_sample_input
-    image_processor = AutoImageProcessor.from_pretrained(model_name)
-    inputs = image_processor(image, return_tensors="pt")
-
-    inputs = inputs["pixel_values"].bfloat16()
-    comments = f"{list(inputs.shape)[-2]}x{list(inputs.shape)[-1]}_batchsize{batch_size}"
-
-    inputs1 = inputs
-    for i in range(batch_size - 1):
-        inputs = torch.cat((inputs, inputs1), dim=0)
-
-    inputs_mesh_mapper = ttnn.ShardTensorToMesh(device, dim=0)
-    weights_mesh_mapper = ttnn.ReplicateTensorToMesh(device)
-    output_mesh_composer = ttnn.ConcatMeshToTensor(device, dim=0)
-
-    torch_resnet50 = load_resnet50_model(model_location_generator)
-    torch_resnet50.eval()
-
-    parameters = preprocess_model_parameters(
-        initialize_model=lambda: torch_resnet50,
-        custom_preprocessor=create_custom_mesh_preprocessor(weights_mesh_mapper),
-        device=None,
-    )
-    torch_resnet50.to(torch.bfloat16)
-
-    test_infra = create_test_infra(
-        device,
-        device_batch_size,
-        model_config["ACTIVATIONS_DTYPE"],
-        model_config["WEIGHTS_DTYPE"],
-        model_config["MATH_FIDELITY"],
-        dealloc_input=True,
-        final_output_mem_config=ttnn.L1_MEMORY_CONFIG,
-        model_location_generator=model_location_generator,
-    )
-    ttnn.synchronize_devices(device)
-
-    num_warmup_iterations = 5
-    num_measurement_iterations = 15
-
-    with torch.no_grad():
-        profiler.start(cpu_key)
-        logits = torch_resnet50(inputs)
-        profiler.end(cpu_key)
-
-        if "resnet50_trace_2cqs" in model_version:
-            run_trace_2cq_model(
-                device,
-                inputs,
-                test_infra,
-                inputs_mesh_mapper,
-                output_mesh_composer,
-                num_warmup_iterations,
-                num_measurement_iterations,
-            )
-        elif "resnet50_2cqs" in model_version:
-            run_2cq_model(
-                device,
-                inputs,
-                test_infra,
-                inputs_mesh_mapper,
-                output_mesh_composer,
-                num_warmup_iterations,
-                num_measurement_iterations,
-            )
-        elif "resnet50_trace" in model_version:
-            run_trace_model(
-                device,
-                inputs,
-                test_infra,
-                inputs_mesh_mapper,
-                output_mesh_composer,
-                num_warmup_iterations,
-                num_measurement_iterations,
-            )
-        elif "resnet50" in model_version:
-            run_model(
-                device,
-                inputs,
-                test_infra,
-                inputs_mesh_mapper,
-                output_mesh_composer,
-                num_warmup_iterations,
-                num_measurement_iterations,
-            )
-        else:
-            assert False, f"Model version to run {model_version} not found"
-
-    first_iter_time = profiler.get(f"compile") + profiler.get(f"cache")
-
-    # ensuring inference time fluctuations is not noise
-    inference_time_avg = profiler.get("run") / num_measurement_iterations
-
-    cpu_time = profiler.get(cpu_key)
-    compile_time = first_iter_time - 2 * inference_time_avg
-    prep_perf_report(
-        model_name=f"ttnn_{model_version}_batch_size{batch_size}",
-        batch_size=batch_size,
-        inference_and_compile_time=first_iter_time,
-        inference_time=inference_time_avg,
-        expected_compile_time=expected_compile_time,
-        expected_inference_time=expected_inference_time,
-        comments=comments,
-        inference_time_cpu=cpu_time,
-    )
-
-    logger.info(
-        f"{model_name} {comments} inference time (avg): {inference_time_avg}, FPS: {batch_size/inference_time_avg}"
-    )
-    logger.info(f"{model_name} compile time: {compile_time}")
