+++ models/demos/ttnn_resnet/tests/perf_e2e_resnet50.py
@@ -10,7 +10,6 @@ import ttnn
 
 from models.utility_functions import (
     profiler,
-    disable_persistent_kernel_cache,
 )
 
 from models.demos.ttnn_resnet.tests.resnet50_test_infra import create_test_infra
@@ -27,34 +26,54 @@ except ModuleNotFoundError:
     use_signpost = False
 
 
+def buffer_address(tensor):
+    addr = []
+    for ten in ttnn.get_device_tensors(tensor):
+        addr.append(ten.buffer_address())
+    return addr
+
+
+def dump_device_profiler(device):
+    if isinstance(device, ttnn.Device):
+        ttnn.DumpDeviceProfiler(device)
+    else:
+        for dev in device.get_device_ids():
+            ttnn.DumpDeviceProfiler(device.get_device(dev))
+
+
+# TODO: Create ttnn apis for this
+ttnn.dump_device_profiler = dump_device_profiler
+
 model_config = {
     "MATH_FIDELITY": ttnn.MathFidelity.LoFi,
     "WEIGHTS_DTYPE": ttnn.bfloat8_b,
     "ACTIVATIONS_DTYPE": ttnn.bfloat8_b,
 }
 
+# TODO: Create ttnn apis for this
+ttnn.buffer_address = buffer_address
+
 
 def run_model(device, tt_inputs, test_infra, num_warmup_iterations, num_measurement_iterations):
-    ops_parallel_config = {}
     tt_inputs_host, input_mem_config = test_infra.setup_l1_sharded_input(device)
     profiler.start("compile")
     test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("compile")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     profiler.start("cache")
     test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("cache")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     for iter in range(0, num_warmup_iterations):
         test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
         _ = ttnn.from_device(test_infra.run(), blocking=True)
-        ttnn.DumpDeviceProfiler(device)
+        ttnn.dump_device_profiler(device)
 
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     if use_signpost:
         signpost(header="start")
     outputs = []
@@ -62,15 +81,14 @@ def run_model(device, tt_inputs, test_infra, num_warmup_iterations, num_measurem
     for iter in range(0, num_measurement_iterations):
         test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
         outputs.append(ttnn.from_device(test_infra.run(), blocking=False))
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     profiler.end(f"run")
     if use_signpost:
         signpost(header="stop")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
 
 def run_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_measurement_iterations):
-    ops_parallel_config = {}
     tt_inputs_host, sharded_mem_config_DRAM, input_mem_config = test_infra.setup_dram_sharded_input(device)
     tt_image_res = tt_inputs_host.to(device, sharded_mem_config_DRAM)
     op_event = ttnn.create_event(device)
@@ -87,7 +105,7 @@ def run_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_meas
     ttnn.record_event(0, op_event)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("compile")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     profiler.start("cache")
     ttnn.wait_for_event(1, op_event)
@@ -98,7 +116,7 @@ def run_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_meas
     ttnn.record_event(0, op_event)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("cache")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     for iter in range(0, num_warmup_iterations):
         ttnn.wait_for_event(1, op_event)
@@ -108,9 +126,9 @@ def run_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_meas
         test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
         ttnn.record_event(0, op_event)
         _ = ttnn.from_device(test_infra.run(), blocking=True)
-        ttnn.DumpDeviceProfiler(device)
+        ttnn.dump_device_profiler(device)
 
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     if use_signpost:
         signpost(header="start")
     outputs = []
@@ -123,17 +141,15 @@ def run_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_meas
         test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
         ttnn.record_event(0, op_event)
         outputs.append(ttnn.from_device(test_infra.run(), blocking=False))
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     profiler.end(f"run")
     if use_signpost:
         signpost(header="stop")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
 
 def run_trace_model(device, tt_inputs, test_infra, num_warmup_iterations, num_measurement_iterations):
-    ops_parallel_config = {}
     tt_inputs_host, input_mem_config = test_infra.setup_l1_sharded_input(device)
-
     # Compile
     profiler.start("compile")
     test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
@@ -142,19 +158,19 @@ def run_trace_model(device, tt_inputs, test_infra, num_warmup_iterations, num_me
     layout = test_infra.input_tensor.layout
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("compile")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
     test_infra.output_tensor.deallocate(force=True)
 
     profiler.start("cache")
     test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("cache")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     # Capture
     test_infra.input_tensor = tt_inputs_host.to(device, input_mem_config)
     test_infra.output_tensor.deallocate(force=True)
-    trace_input_addr = test_infra.input_tensor.buffer_address()
+    trace_input_addr = ttnn.buffer_address(test_infra.input_tensor)
     tid = ttnn.begin_trace_capture(device, cq_id=0)
     tt_output_res = test_infra.run()
     tt_image_res = ttnn.allocate_tensor_on_device(
@@ -165,16 +181,16 @@ def run_trace_model(device, tt_inputs, test_infra, num_warmup_iterations, num_me
         input_mem_config,
     )
     ttnn.end_trace_capture(device, tid, cq_id=0)
-    assert trace_input_addr == tt_image_res.buffer_address()
-    ttnn.DumpDeviceProfiler(device)
+    assert trace_input_addr == ttnn.buffer_address(tt_image_res)
+    ttnn.dump_device_profiler(device)
 
     for iter in range(0, num_warmup_iterations):
         ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res)
         ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
         _ = ttnn.from_device(tt_output_res, blocking=True)
-        ttnn.DumpDeviceProfiler(device)
+        ttnn.dump_device_profiler(device)
 
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     if use_signpost:
         signpost(header="start")
     outputs = []
@@ -183,17 +199,16 @@ def run_trace_model(device, tt_inputs, test_infra, num_warmup_iterations, num_me
         ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res)
         ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
         outputs.append(ttnn.from_device(tt_output_res, blocking=False))
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     profiler.end(f"run")
     if use_signpost:
         signpost(header="stop")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     ttnn.release_trace(device, tid)
 
 
 def run_trace_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, num_measurement_iterations):
-    ops_parallel_config = {}
     tt_inputs_host, sharded_mem_config_DRAM, input_mem_config = test_infra.setup_dram_sharded_input(device)
     tt_image_res = tt_inputs_host.to(device, sharded_mem_config_DRAM)
 
@@ -214,7 +229,7 @@ def run_trace_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, nu
     ttnn.record_event(0, op_event)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("compile")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     profiler.start("cache")
     ttnn.wait_for_event(1, op_event)
@@ -228,21 +243,21 @@ def run_trace_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, nu
     test_infra.output_tensor.deallocate(force=True)
     _ = ttnn.from_device(test_infra.run(), blocking=True)
     profiler.end("cache")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     # Capture
     ttnn.wait_for_event(1, op_event)
     ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
     ttnn.record_event(1, write_event)
-
     ttnn.wait_for_event(0, write_event)
     test_infra.input_tensor = ttnn.to_memory_config(tt_image_res, input_mem_config)
     ttnn.record_event(0, op_event)
     test_infra.output_tensor.deallocate(force=True)
-    trace_input_addr = test_infra.input_tensor.buffer_address()
+    trace_input_addr = ttnn.buffer_address(test_infra.input_tensor)
+
     tid = ttnn.begin_trace_capture(device, cq_id=0)
     tt_output_res = test_infra.run()
-    reshard_out = ttnn.allocate_tensor_on_device(
+    input_tensor = ttnn.allocate_tensor_on_device(
         shape,
         dtype,
         layout,
@@ -250,20 +265,20 @@ def run_trace_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, nu
         input_mem_config,
     )
     ttnn.end_trace_capture(device, tid, cq_id=0)
-    assert trace_input_addr == reshard_out.buffer_address()
-    ttnn.DumpDeviceProfiler(device)
+    assert trace_input_addr == ttnn.buffer_address(input_tensor)
+    ttnn.dump_device_profiler(device)
 
     for iter in range(0, num_warmup_iterations):
         ttnn.wait_for_event(1, op_event)
         ttnn.copy_host_to_device_tensor(tt_inputs_host, tt_image_res, 1)
         ttnn.record_event(1, write_event)
         ttnn.wait_for_event(0, write_event)
-        reshard_out = ttnn.reshard(tt_image_res, input_mem_config, reshard_out)
+        input_tensor = ttnn.reshard(tt_image_res, input_mem_config, input_tensor)
         ttnn.record_event(0, op_event)
         ttnn.execute_trace(device, tid, cq_id=0, blocking=True)
-        ttnn.DumpDeviceProfiler(device)
+        ttnn.dump_device_profiler(device)
 
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     if use_signpost:
         signpost(header="start")
     outputs = []
@@ -274,21 +289,21 @@ def run_trace_2cq_model(device, tt_inputs, test_infra, num_warmup_iterations, nu
         ttnn.record_event(1, write_event)
         ttnn.wait_for_event(0, write_event)
         # TODO: Add in place support to ttnn to_memory_config
-        reshard_out = ttnn.reshard(tt_image_res, input_mem_config, reshard_out)
+        input_tensor = ttnn.reshard(tt_image_res, input_mem_config, input_tensor)
         ttnn.record_event(0, op_event)
         ttnn.execute_trace(device, tid, cq_id=0, blocking=False)
         outputs.append(tt_output_res.cpu(blocking=False))
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
     profiler.end(f"run")
     if use_signpost:
         signpost(header="stop")
-    ttnn.DumpDeviceProfiler(device)
+    ttnn.dump_device_profiler(device)
 
     ttnn.release_trace(device, tid)
 
 
 def run_perf_resnet(
-    batch_size,
+    device_batch_size,
     expected_inference_time,
     expected_compile_time,
     hf_cat_image_sample_input,
@@ -297,9 +312,12 @@ def run_perf_resnet(
     model_location_generator,
 ):
     profiler.clear()
-    disable_persistent_kernel_cache()
-    if batch_size <= 2:
+    if device_batch_size <= 2:
         pytest.skip("Batch size 1 and 2 are not supported with sharded data")
+
+    is_mesh_device = isinstance(device, ttnn.MeshDevice)
+    num_devices = device.get_num_devices() if is_mesh_device else 1
+    batch_size = device_batch_size * num_devices
     first_key = f"first_iter_batchsize{batch_size}"
     second_key = f"second_iter_batchsize{batch_size}"
     cpu_key = f"ref_key_batchsize{batch_size}"
@@ -323,7 +341,7 @@ def run_perf_resnet(
 
     test_infra = create_test_infra(
         device,
-        batch_size,
+        device_batch_size,
         model_config["ACTIVATIONS_DTYPE"],
         model_config["WEIGHTS_DTYPE"],
         model_config["MATH_FIDELITY"],
@@ -331,8 +349,7 @@ def run_perf_resnet(
         final_output_mem_config=ttnn.L1_MEMORY_CONFIG,
         model_location_generator=model_location_generator,
     )
-
-    ttnn.synchronize_device(device)
+    ttnn.synchronize_devices(device)
 
     num_warmup_iterations = 5
     num_measurement_iterations = 15
@@ -342,7 +359,6 @@ def run_perf_resnet(
         logits = torch_resnet50(inputs)
         profiler.end(cpu_key)
 
-        # tt_inputs = tt_resnet50.preprocessing(inputs)
         if "resnet50_trace_2cqs" in model_version:
             run_trace_2cq_model(device, inputs, test_infra, num_warmup_iterations, num_measurement_iterations)
         elif "resnet50_2cqs" in model_version:
